Feature/permissions removal

See merge request grafolean/grafolean!24

+++ backend/datatypes.py
@@ -627,7 +627,8 @@ class Permission(object):
             else:
                 c.execute('SELECT id, user_id, resource_prefix, methods FROM permissions WHERE user_id = %s ORDER BY resource_prefix, id;', (user_id,))
             for permission_id, user_id, resource_prefix, methods in c:
-                ret.append({'id': permission_id, 'user_id': user_id, 'resource_prefix': resource_prefix, 'methods': methods})
+                methods_as_list = None if not methods else [m.strip('{ }') for m in methods.split(',')]  # not sure why, but we get what we inserted (string instead of a list)... this is workaround
+                ret.append({'id': permission_id, 'user_id': user_id, 'resource_prefix': resource_prefix, 'methods': methods_as_list})
             return ret
 
     def insert(self):
@@ -658,8 +659,10 @@ class Permission(object):
     @staticmethod
     def delete(permission_id):
         with db.cursor() as c:
-            c.execute('DELETE FROM permissions WHERE id = %s;', (permission_id,))
-            return c.rowcount
+            c.execute('DELETE FROM permissions WHERE id = %s RETURNING user_id;', (permission_id,))
+            num_deleted = c.rowcount
+            user_id = c.fetchone()[0]
+            return num_deleted, user_id
 
 
 class Bot(object):
